requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.1
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-js-asset==1.2.2
djangorestframework==3.11.1
gunicorn==20.0.4
Markdown==3.2.2
Pillow==7.2.0
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
django-os-geocoder~=0.1.6